v1.6.2 Update || New Black List Permission Option
Hello, I have recently gotten a request to add a blacklist permission
option to Vouchers. What this allows you to do is check if a player has
1 or more permissions that might be blacklisted. If they have one of
those permissions then they will not be able to use the voucher.

Added:
  - New blacklist permission option. This allows you to make the player
not able to use a voucher if they have a permission.

Changes:
  - Changed the packages names so they follow the java standerds.
  - Version class was moved to the API package.
  - Some small code clean up to make it nicer.